+++ redundancy_payments_alpha/claimants_user_journey/forms/claimant_contact_details.py
@@ -1,16 +1,9 @@
 import re
-from datetime import date
 from flask_wtf import Form
 from wtforms import TextField, SelectField, StringField, ValidationError
 from wtforms.fields.html5 import TelField, EmailField
 from wtforms.validators import DataRequired, Length, Email, AnyOf, Regexp
-
-
-def convert_string_to_date(date_string):
-    split_string = date_string.split('/')
-    if type(split_string) != list or len(split_string) != 3:
-        raise SyntaxError('Was expecting dd/MM/yyyy.')
-    return date(int(split_string[2]), int(split_string[1]), int(split_string[0]))
+from validators.Validators import DateOfBirthValidator
 
 
 class ClaimantContactDetails(Form):
@@ -55,16 +48,5 @@ class ClaimantContactDetails(Form):
                                  Regexp(regex=re.compile('^[A-Z]{2}[0-9]{6}[A-Z]{1}$'),
                                         message="National Insurance Number must be two letters followed by six digits and a further letter (e.g. 'AB123456C').")])
     date_of_birth = TextField('Date Of Birth',
-                              validators=[DataRequired(),
-                                          Regexp(regex=re.compile('^[0-9]{2}[/][0-9]{2}[/][0-9]{4}$'),
-                                                 message="Date Of Birth must be in the format dd/mm/yyyy.") ])
-
-    def validate_date_of_birth(form, field):
-        date_of_birth = form._fields.get('date_of_birth')
-        try:
-            parsed_date = convert_string_to_date(date_of_birth.data)
-        except SyntaxError:
-            raise ValidationError('Date Of Birth must be greater than or equal to 1900 and not in the future.')
-        if parsed_date.year < 1900 or parsed_date >= date.today():
-            raise ValidationError('Date Of Birth must be greater than or equal to 1900 and not in the future.')
+                              validators=[DataRequired(), DateOfBirthValidator(format_message="Date Of Birth must be in the format dd/mm/yyyy.")])
