+++ redundancy_payments_alpha/unit_tests/claimant_contact_details/test_form_validation.py
@@ -334,13 +334,3 @@ class TestDateOfBirthValidation(unittest.TestCase):
         # then
         assert_that(form.date_of_birth.errors, has_item("Date Of Birth must be greater than or equal to 1900 and not in the future.") )
 
-    def test_date_of_birth_field_does_not_display_int_parsing_error(self):
-        # given
-        entered_date = complete_form_data()
-        entered_date['date_of_birth'] = 'aa/bb/cccc'
-        # when
-        form = complete_form(entered_date)
-        form.validate()
-        # then
-        assert_that(form.date_of_birth.errors, not(has_item("invalid literal for int() with base 10: 'cccc'")))
-        assert_that(form.date_of_birth.errors, has_item("invalid literal for int() with base 10: 'cccc'"))
